requirements.txt
python-dotenv
transformers
torch